Restore default led behavior
With the recent template, when the power is turned on, the led is lighting up blue, which is pretty annoying  especially when it's getting dark.

With the new template, if the plug is off, no light lighting up and if its turned on, a pretty weak red light is shining. With this it's no problem to use it in the sleeping room.