Update gosund-SW2
I had issues with blinking when going over 85%. I adjusted the range and then ran into an issue where if you set the dimmer to 0 it would come back on; moving line 52 up into the previous statement fixed it. I've tested in the web interface, web console and Home Assistant. I also tested with an incandescent and a dimable led bulb.