Change GPIO0 from Switch1 (9) to Button1 (17)
When using Tasmota 8.1.0 at least and just applying the template with Switch1 (9) the device when physically pressed will turn on and off again immediately. Other commands can be added in the Tasmota console to make the device work with Switch1. But changing to a Button1 (17) in the template will allow this device to work by just applying the template. Using Button1 will allow the physical device to toggle on and off with each physical press.