Update aoycocr_SW1
Current template is incorrect. It does not behave like the description (with Tasmota 8.1.0.x). It actually acts like the red LED is an inverted LedLink, and the blue LED is on when the relay is off (i.e. inverted relay state). The end result is that it looks like purple when off, and red when on).

This change implements LedLink, and properly inverts all LEDs.

Unfortunately there's no nice way to use the blue LED with the current Tasmota implementation, but it's there as Led2i, if someone wants to use it in rules (and accept the consequences).