Update treatlife_SS02
URL points to a newer version of this switch which cannot be flash with tuya-convert (not the ESP chip) which I found out the hard way.  The link I put in goes to the older version of the switch.  The replaced link should probably go with a warning in the incompatible devices section.