Update moes-MS-104B
I'm 100% sure that the counters are sleep sensitive (at least in 6.6.0.13) . Propose change for SwitchMode1 is tested for more than 2 months with two devices. It's too risky for me to do an upgrade to 8.1 but if it works for 6.6.0.13 should work also for 8.1. Without that changes the script heavily bouncing.