Update hugoai_awp02l-n
Assuming that this is the same plug as the HugoAI version, I would recommend a slight change to the template.

At least on the HugoAi version, setting GPIO00 to LED2i (57) prevents the power LED from turning on when the plug is toggled on.  Setting GPIO00 to LED3i (58) instead, gives you both a power LED when toggled on and the MQTT heartbeat blink when there is no connection to the MQTT server (but only when the plug is in the off state).