Merge two relays into one
This device only has one outlet, and the previous settings created two relays, as a result, the MQTT RESULT contains POWER1 instead of POWER. There is no need to be able to control each relay independently. By setting both GPIO4 and GPIO5 to `relay1` it removes the need for the "power2 on boot hack" and makes the MQTT result cleaner. Therefore this will appear as a single relay device, without having to deal with POWER1 / POWER2.

Also fixed typo in the template name