Add IR Remote support to config
These bulbs come with an IR Remote as seen here: 
https://www.novostella.net/products/2-pack-9w-wi-fi-rf-control-rgbcw-smart-light-bulbs 

I added the GPIO PIN for the IR receiver and made a rule to get all basic features of the remote working. 
There's three buttons not working yet as these would require custom builds with expressions enabled and possibly even scripting. Might add an extended rule in the future. 
Previous change request did not include the GPIO Pin and I improved the ruleset