Update hombli_duo
With 15 values in the gpio array, the html page gets an extra "GPIO014". As entering the malformed template in Tasmota removed the extra last pin of the array, I'm assuming that also removing this from the template works no worse than leaving it there, but not having the device I'm not able to verify if it works, before or after this change.