Alternate Gosund SW5 Template w/ red 'ON' LED
The way these outlets come from the factory- or at least the ones I received- are completely dark when 'OFF' and show a red LED when 'ON'.  The current template here shows a bright green LED instead.  I couldn't find a template that used the red LED, so I cooked up the `template9_alt` one above to do just that