GD-30W: update RGB rule and water sensor Rule1 is updated to make RGB cycle switch work again. Rule2 allows to publish water sensor to MQTT as the previously used error switch looks like not working anymore. The Home §Assistant integration configuration is also updated to update the water status binary sensor. All credits go to the international Home Assistant community forum members who have provided all these updated rules.
The final version:
* Rules are better separated;
* Use POWER4 instead of a dedicated MQTT topic for the water sensor and use it also in the Home Assistant config thanks to @SirGoodenough
Tried to keep previous values for light mode but definitly doe not work for me, looks like some devices are a little bit different.

Adjust code to be consistent with these updated durring this week for water sensor and add RGB/color mode activation code which works for me.