Update gosund_SW6 (#1160)
* Update gosund_SW6

The current (non-explicit) configuration of gosund_SW6 worked when using toggle commands.  The problem comes when trying to control a group.  If an off command was sent but, the gosund was already off it would toggle to on.

The current (explicit) configuration had similar problem in that if the relay was off but the circuit was on nothing would change when receiving an off command. This means the switch would only take external commands if relay2's state matched the current state of the circuit.  

This new change adds another relay which will change to match external commands. It is also kept in sync with the state of the circuit so that if an OFF command was received nothing happens.  Using relay1 won't work because it will cause a loop in certain circumstances since it is set to match the status of the circuit.

* Update gosund_SW6

* Update gosund_SW6