Just 1 or 2 suggestions
Most important one is line 80, where the rules for battery status reporting is being setup, the status is published to "BATT" however in line 88 the full word "BATTERY" is used. This cause Home Assistant to report the battery status as "Unknown". AS soon as I updated the console command in line 80 it started to show up properly... I guess I could be changed in either line just the both should be the same.
 
Secondly, might be nothing, I explicitly added instructions to enable rule2

Finally, I didn't change this since mine worked as expected following the instructions. Line 66 set "dpId" to 51, however in the output for validation on line 71 it is set as 21. I couldn't find anything on the TuyaMCU documentation pages to confirm what the value should be so I just kept mine as 51. As I said, it's working for me so 21 in line 71 might be just an oversight