Update mirabella_genio_I002899
The relay GPIO for this device is 14. After programming with Tasmota I used the template that was on the site and the relay did not work. After following the tracks on the PCB I found that the relay GPIO is 14. In fact the TYWE2S does not even have a GPIO 15 pin.
I have corrected the template.
I have also created a blog post with how to open this device. You may want to link to it? 
https://homeopen.wordpress.com/2022/04/27/mirabella-genio-power-plug-two-usb/