Update arlec_SGS01HA (#1478)
This adds support for more than 10 devices. The sensor number comes as a 2 digit hexadecimal number, so when you get to 0A it previously just parsing it to 0, so you would have multiple devices with ID of 0. Bunnings have these cheap on clearance so I have more than 10 devices now.