Disable stemming in search
We're always performing wildcard searches and for those, stemming is not
run at search time, so running it at index time leads to unexpected
results. So this commit removes it from the index pipeline.

It also removes it from the search pipeline, just in case we ever remove
the wildcard.