Search fixes (#1561)
* Disable stemming in search

We're always performing wildcard searches and for those, stemming is not
run at search time, so running it at index time leads to unexpected
results. So this commit removes it from the index pipeline.

It also removes it from the search pipeline, just in case we ever remove
the wildcard.

* Apply wildcard to all search tokens

Previously for a search term consisting of multiple words, the search
was left-anchored only for the last word, all other would be search as
an exact match. This commit changes that by moving the logic to the
search pipeline an applying the wildcard to every token.