Update kogan-KSH2MLEDSPA
I own this product and use it with ESPHome, and have figured out GPIO0 pin is connected to the physical button on the device (and mapped it to toggle the light like in the original firmware). Using the old template codes, I have added it, hopefully can help someone who wants to know what GPIO the button is.