Update wyze_WLPPO1 - Add information about Roku Outdoor Smart Plug SE (identical hardware) (#1951)
The Roku Outdoor Smart Plug SE has identical hardware and FCC ID, but it has all of the security features buttoned up properly. I'm pretty sure it's not possible to flash this with custom firmware with current knowledge.

I assume the same is true of the corresponding indoor 1-plug Roku smartplug, which is also a rebranding of Wyze hardware, but I do not have this hardware to check.